backend/requirements.txt
git+git://github.com/Edinburgh-Genome-Foundry/DnaCauldron.git
git+git://github.com/Edinburgh-Genome-Foundry/DnaFeaturesViewer
git+git://github.com/Edinburgh-Genome-Foundry/DnaChisel
git+git://github.com/Edinburgh-Genome-Foundry/Caravagene.git
git+git://github.com/Edinburgh-Genome-Foundry/bandwitch.git
git+git://github.com/Edinburgh-Genome-Foundry/Proglog

git+https://github.com/pygraphviz/pygraphviz.git

redis==2.10.5
rq==0.6.0
requests==2.11.1
django==1.10.5
djangorestframework==3.5.3
drfdocs==0.0.11
drf-tracking==1.0.1
django-cors-headers==2.0.0
django_rq==0.9.2
flametree==0.1.5a
WeasyPrint==0.34
gunicorn~=19.3
watchdog==0.8.3
bandwagon==0.1.1
jinja2==2.9
networkx==2.0
